Improve UI of advanced commit options
